Add script to enable RPMFusion

+++ brace-fedora-enable-auto-updates.sh
@@ -4,5 +4,10 @@
 
 if [[ $EUID -ne 0 ]]; then echo 'This script needs to be run as root!'; exit 1; fi;
 
-dnf install dnf-automatic;
-systemctl enable dnf-automatic-install.timer --now;
+echo "Enabling automatic update download/install";
+if [ -f /etc/fedora-release ]; then
+	dnf install dnf-automatic;
+	systemctl enable dnf-automatic-install.timer --now;
+else
+	echo "Currently only Fedora is supported";
+fi;
